requirements.txt
Django==3.1.3
django-crispy-forms==1.10.0
gunicorn==20.0.4
idna==2.10
image==1.5.33
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
Pillow==8.0.1
pylint==2.6.0
python-dateutil==1.5
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1